requirements.txt
psycopg2==2.9.1
Django==3.0.1
django-heroku==0.3.1
gunicorn
numpy
scipy
opencv-python
Pillow
imageio

